Face-lift with VuePress
Some huge changes here that I have been working on for a while,
It's still a work in progress but some major changes have occured.

The following changes have been made:

- A full converstion of the site to VuePress *excited*
- Addition of a blog featue - not sure how much this will be used
- Full .md support, letting me create pages on the fly

Still not sure how much I will be using all of this added
functionality, if this amounts to nothing then at least it's
something new that i have learned.

- Making a custom VuePress theme
- Creating plugins
- Optimising CSS turning 1000+ lines into a few hundered